+++ menelaus/experimental/detector.py
@@ -1,5 +1,7 @@
+from toolz import pipe
+
 class Detector():
-    def __init__(self, representation, alarm, probes : list = None):
+    def __init__(self, representation, alarm, operators: list = None):
         self.representation = representation
         self.alarm = alarm
         # summary statistics, live plots, etc.
@@ -12,19 +14,8 @@ class Detector():
         raise NotImplementedError
 
     def run(self):
-        pass
-
-    def recalibrate(self, rep_test):
-        if self.state == "drift":
-            if len(rep_test) == 1 and window_not_full:
-                # TODO store until window size exceeded, then merge
-                # TODO representation should be in charge of knowing
-                #      how to add one / many representations to itself
-                self._rep_pool.append(rep_test)
-            elif len(rep_test) == 1 and window_full:
-                self.rep_reference = self.rep_reference + self._rep_pool
-            else:
-                self.rep_reference = self.rep_reference + rep_test
+        # ...
+        self = pipe(self, *self.probes)
 
     @property
     def state(self):
@@ -43,4 +34,4 @@ class BasicDetector(Detector):
             self.alarm.evaluate(self.rep_reference, self.rep_test)
 
             if not initial_rep_test:
-                self.recalibrate(self.rep_test)
+                self.representation.recalibrate(data)
